requirements.txt
elevenlabs==0.2.27
python-dotenv==1.0.0
click==8.1.7
google-generativeai==0.3.0
sounddevice==0.4.6
numpy>=1.20.0
soundfile>=0.12.1
